Remove unnecessary stub and files.
